+++ src/main/java/com/flightstats/hub/app/AwsBindings.java
@@ -5,8 +5,8 @@ import com.amazonaws.services.s3.AmazonS3;
 import com.flightstats.hub.cluster.CuratorCluster;
 import com.flightstats.hub.dao.*;
 import com.flightstats.hub.dao.aws.*;
-import com.flightstats.hub.group.GroupDao;
 import com.flightstats.hub.spoke.*;
+import com.flightstats.hub.webhook.GroupDao;
 import com.google.inject.AbstractModule;
 import com.google.inject.Inject;
 import com.google.inject.Provides;
