+++ src/main/java/com/flightstats/hub/app/HubBindings.java
@@ -7,9 +7,6 @@ import com.fasterxml.jackson.databind.module.SimpleModule;
 import com.flightstats.hub.alert.AlertRunner;
 import com.flightstats.hub.channel.ChannelValidator;
 import com.flightstats.hub.cluster.*;
-import com.flightstats.hub.group.GroupProcessor;
-import com.flightstats.hub.group.GroupProcessorImpl;
-import com.flightstats.hub.group.GroupValidator;
 import com.flightstats.hub.health.HubHealthCheck;
 import com.flightstats.hub.metrics.HostedGraphiteSender;
 import com.flightstats.hub.metrics.MetricsRunner;
@@ -24,6 +21,9 @@ import com.flightstats.hub.spoke.GCRunner;
 import com.flightstats.hub.time.NtpMonitor;
 import com.flightstats.hub.time.TimeService;
 import com.flightstats.hub.util.HubUtils;
+import com.flightstats.hub.webhook.GroupProcessor;
+import com.flightstats.hub.webhook.GroupProcessorImpl;
+import com.flightstats.hub.webhook.GroupValidator;
 import com.google.inject.AbstractModule;
 import com.google.inject.Provides;
 import com.google.inject.Singleton;
