+++ src/main/java/com/flightstats/hub/dao/aws/DynamoGroupDao.java
@@ -3,8 +3,8 @@ package com.flightstats.hub.dao.aws;
 import com.amazonaws.services.dynamodbv2.AmazonDynamoDBClient;
 import com.amazonaws.services.dynamodbv2.model.*;
 import com.flightstats.hub.app.HubServices;
-import com.flightstats.hub.group.Group;
-import com.flightstats.hub.group.GroupDao;
+import com.flightstats.hub.webhook.Group;
+import com.flightstats.hub.webhook.GroupDao;
 import com.google.common.base.Optional;
 import com.google.common.util.concurrent.AbstractIdleService;
 import com.google.inject.Inject;
