+++ src/main/java/com/flightstats/hub/dao/aws/S3BatchManager.java
@@ -2,11 +2,11 @@ package com.flightstats.hub.dao.aws;
 
 import com.flightstats.hub.app.HubServices;
 import com.flightstats.hub.dao.ChannelService;
-import com.flightstats.hub.group.Group;
-import com.flightstats.hub.group.GroupService;
 import com.flightstats.hub.model.ChannelConfig;
 import com.flightstats.hub.replication.S3Batch;
 import com.flightstats.hub.util.HubUtils;
+import com.flightstats.hub.webhook.Group;
+import com.flightstats.hub.webhook.GroupService;
 import com.google.common.util.concurrent.AbstractIdleService;
 import com.google.inject.Inject;
 import org.slf4j.Logger;
