+++ src/main/java/com/flightstats/hub/replication/S3Batch.java
@@ -1,9 +1,9 @@
 package com.flightstats.hub.replication;
 
 import com.flightstats.hub.app.HubProperties;
-import com.flightstats.hub.group.Group;
 import com.flightstats.hub.model.ChannelConfig;
 import com.flightstats.hub.util.HubUtils;
+import com.flightstats.hub.webhook.Group;
 import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
