+++ src/main/java/com/flightstats/hub/util/HubUtils.java
@@ -6,8 +6,8 @@ import com.fasterxml.jackson.databind.node.ObjectNode;
 import com.flightstats.hub.app.HubHost;
 import com.flightstats.hub.app.HubProvider;
 import com.flightstats.hub.cluster.CuratorCluster;
-import com.flightstats.hub.group.Group;
 import com.flightstats.hub.model.*;
+import com.flightstats.hub.webhook.Group;
 import com.google.common.base.Optional;
 import com.google.common.io.ByteStreams;
 import com.google.inject.Inject;
