+++ src/main/java/com/flightstats/hub/webhook/GroupContentPathSet.java
@@ -1,4 +1,4 @@
-package com.flightstats.hub.group;
+package com.flightstats.hub.webhook;
 
 import com.flightstats.hub.model.ContentPath;
 import com.google.inject.Inject;
