+++ src/main/java/com/flightstats/hub/webhook/GroupDao.java
@@ -1,4 +1,4 @@
-package com.flightstats.hub.group;
+package com.flightstats.hub.webhook;
 
 import com.google.common.base.Optional;
