+++ src/main/java/com/flightstats/hub/webhook/GroupProcessor.java
@@ -1,4 +1,4 @@
-package com.flightstats.hub.group;
+package com.flightstats.hub.webhook;
 
 public interface GroupProcessor {
