+++ src/main/java/com/flightstats/hub/webhook/GroupValidator.java
@@ -1,4 +1,4 @@
-package com.flightstats.hub.group;
+package com.flightstats.hub.webhook;
 
 import com.flightstats.hub.channel.ChannelValidator;
 import com.flightstats.hub.exception.InvalidRequestException;
