+++ src/main/java/com/flightstats/hub/webhook/QueryGenerator.java
@@ -1,4 +1,4 @@
-package com.flightstats.hub.group;
+package com.flightstats.hub.webhook;
 
 import com.flightstats.hub.model.Location;
 import com.flightstats.hub.model.TimeQuery;
