+++ src/test/java/com/flightstats/hub/webhook/GroupErrorTest.java
@@ -1,4 +1,4 @@
-package com.flightstats.hub.group;
+package com.flightstats.hub.webhook;
 
 import com.flightstats.hub.test.Integration;
 import org.apache.curator.framework.CuratorFramework;
