+++ src/test/java/com/flightstats/hub/webhook/GroupValidatorTest.java
@@ -1,4 +1,4 @@
-package com.flightstats.hub.group;
+package com.flightstats.hub.webhook;
 
 import com.flightstats.hub.exception.InvalidRequestException;
 import com.google.common.base.Strings;
